Minor fixes in dispatcher and moving sanitizer to pkg
